requirements.txt
pywebview==5.4
pyserial==3.5